# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI-powered E-Commerce Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /support/chat:
    post:
      operationId: supportChat
      tags: [support]
      summary: Chat with customer support AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupportChatBody"
      responses:
        "200":
          description: Support response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportChatResponse"
  /support/return:
    post:
      operationId: initiateReturn
      tags: [support]
      summary: Initiate a return request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiateReturnBody"
      responses:
        "200":
          description: Return initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnResponse"
...
components:
  schemas:
...
    SupportChatBody:
      type: object
      properties:
        message:
          type: string
        sessionId:
          type: string
        orderId:
          type: integer
        conversationHistory:
          type: array
          items:
            type: object
            properties:
              role:
                type: string
              content:
                type: string
            required: [role, content]
      required: [message, sessionId]
    SupportChatResponse:
      type: object
      properties:
        response:
          type: string
        intent:
          type: string
        suggestedActions:
          type: array
          items:
            type: string
        escalate:
          type: boolean
      required: [response, intent, escalate]
    InitiateReturnBody:
      type: object
      properties:
        orderId:
          type: integer
        sessionId:
          type: string
        reason:
          type: string
        items:
          type: array
          items:
            type: integer
      required: [orderId, sessionId, reason]
    ReturnResponse:
      type: object
      properties:
        returnId:
          type: string
        status:
          type: string
        message:
          type: string
        estimatedRefund:
          type: number
        instructions:
          type: string
      required: [returnId, status, message, instructions]
...